Remove LazyExtensions
